Allow passing a platform argument to generate graph for CLJS
This required an upgrade to tools.namespace too. The new version of
tools.namespace requires tools.reader 0.10.0, but when I was testing
this there was another version of tools.reader coming from somewhere
which was conflicting. I've opened a ticket against Leiningen to upgrade
their clj-http dependency as I suspect it could be this one.

This patch had special help from @atroche in fixing merge conflicts.
Thanks!